+++ upload/library/Iversia/FAQ/Listener/TemplateHook.php
@@ -24,7 +24,7 @@ class Iversia_FAQ_Listener_TemplateHook
             $faqLinks = array();
             $faqLinks['faqCats'] = XenForo_Application::getSimpleCacheData('faq_categories');
 
-            if (! $faqLinks['faqCats']) {
+            if (! isset($faqLinks['faqCats'])) {
                 // Create cache
                 $categories = $faqCatModel->getAll(XenForo_Application::get('options')->faqCatNavCount);
                 $faqLinks['faqCats'] = XenForo_Application::setSimpleCacheData('faq_categories', $categories);